server/requirements.txt
fastapi == 0.63.0
web3==5.16.0
uvicorn==0.13.3
py-solc-x==1.1.0